refactor: dealer의 comparePlayer를 player의 calculatePlayerWinningStatus로 리팩터링
- 기존 메서드에서는 dealer의 승무패인지, player의 승무패인지가 명확하지 않았음.
- 메서드를 실행하는 주체를 player로 바꾸고 메서드의 이름을 변경해 player의 승무패라는 것을 명시
