+++ src/test/java/blackjack/domain/user/PlayerTest.java
@@ -1,43 +1,41 @@
 package blackjack.domain.user;
 
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.SoftAssertions.assertSoftly;
+
 import blackjack.domain.card.Card;
 import blackjack.domain.card.CardGroup;
 import blackjack.domain.card.CardNumber;
 import blackjack.domain.card.CardShape;
 import blackjack.domain.card.Deck;
 import blackjack.domain.card.TestDeckGenerator;
-import blackjack.domain.user.Player;
-import blackjack.domain.user.User;
+import blackjack.domain.result.WinningStatus;
+import java.util.List;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
-
-import java.util.List;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.SoftAssertions.assertSoftly;
 
 class PlayerTest {
 
     private final String name = "test";
-    private final Card firstCard = new Card(CardShape.SPADE, CardNumber.ACE);
-    private final Card secondCard = new Card(CardShape.CLOVER, CardNumber.EIGHT);
+    private final Card cardKing = new Card(CardShape.SPADE, CardNumber.KING);
+    private final Card cardEight = new Card(CardShape.CLOVER, CardNumber.EIGHT);
     private CardGroup initialGroup;
 
     @BeforeEach
     void setUp() {
-        initialGroup = new CardGroup(firstCard, secondCard);
+        initialGroup = new CardGroup(cardKing, cardEight);
     }
 
     @Test
     @DisplayName("플레이어 초기화 테스트")
     void initTest() {
         final User player = new Player(name, initialGroup);
-
         assertSoftly(softly -> {
             softly.assertThat(player.getName()).isEqualTo(name);
-            softly.assertThat(player.getCardGroups().getCards()).containsExactly(firstCard, secondCard);
+            softly.assertThat(player.getCardGroups().getCards()).containsExactly(cardKing, cardEight);
         });
     }
 
@@ -47,7 +45,7 @@ class PlayerTest {
         final User player = new Player(name, initialGroup);
 
         assertThat(player.getScore().getValue())
-                .isEqualTo(firstCard.getNumber().getValue() + secondCard.getNumber().getValue());
+                .isEqualTo(cardKing.getNumber().getValue() + cardEight.getNumber().getValue());
     }
 
     @Test
@@ -59,7 +57,7 @@ class PlayerTest {
 
         player.drawCard(deck);
 
-        Assertions.assertThat(player.getCardGroups().getCards()).containsExactly(firstCard, secondCard, card);
+        Assertions.assertThat(player.getCardGroups().getCards()).containsExactly(cardKing, cardEight, card);
     }
 
     @Test
@@ -67,6 +65,72 @@ class PlayerTest {
     void getInitialStatus() {
         final User player = new Player(name, initialGroup);
 
-        assertThat(player.getFirstOpenCardGroup().getCards()).containsExactly(firstCard, secondCard);
+        assertThat(player.getFirstOpenCardGroup().getCards()).containsExactly(cardKing, cardEight);
+    }
+
+    @Nested
+    @DisplayName("딜러와 플레이어의 카드를 비교하여, WinningStatus를 반환하는 기능 테스트")
+    class comparePlayerTest {
+
+        @Test
+        @DisplayName("딜러가 bust 플레이어가 bust일 때 패배가 반환")
+        void comparePlayerTestIfDealerBustPlayerBust() {
+            final CardGroup bustCardGroup = initialGroup;
+            bustCardGroup.add(new Card(CardShape.CLOVER, CardNumber.FIVE));
+            final Dealer dealer = new Dealer(bustCardGroup);
+            final Player player = new Player("제이미", bustCardGroup);
+
+            assertThat(player.calculatePlayerWinningStatus(dealer)).isEqualTo(WinningStatus.LOSE);
+        }
+
+        @Test
+        @DisplayName("딜러가 bust 플레이어가 bust가 아닐 때 승리가 반환")
+        void comparePlayerTestIfDealerBustPlayerNonBust() {
+            final CardGroup bustCardGroup = new CardGroup(cardKing, cardEight);
+            bustCardGroup.add(new Card(CardShape.CLOVER, CardNumber.FIVE));
+            final Dealer dealer = new Dealer(bustCardGroup);
+            final Player player = new Player("제이미", initialGroup);
+
+            assertThat(player.calculatePlayerWinningStatus(dealer)).isEqualTo(WinningStatus.WIN);
+        }
+
+        @Test
+        @DisplayName("딜러와 플레이어가 bust가 아니고, score가 동일할 때 무승부 반환")
+        void comparePlayerEqualScore() {
+            final Dealer dealer = new Dealer(new CardGroup(cardKing, cardEight));
+            final Player player = new Player("필립", initialGroup);
+
+            assertThat(player.calculatePlayerWinningStatus(dealer)).isEqualTo(WinningStatus.TIE);
+        }
+
+        @Test
+        @DisplayName("딜러와 플레이어가 bust가 아니고, 플레이어의 스코어가 더 클 때 승리 반환")
+        void comparePlayerIfPlayerScoreBiggerThanDealer() {
+            final Dealer dealer = new Dealer(new CardGroup
+                    (new Card(CardShape.HEART, CardNumber.FIVE), new Card(CardShape.CLOVER, CardNumber.EIGHT)));
+            final Player player = new Player("제이미", initialGroup);
+
+            assertThat(player.calculatePlayerWinningStatus(dealer)).isEqualTo(WinningStatus.WIN);
+        }
+
+        @Test
+        @DisplayName("딜러와 플레이어는 버스트가 아니고 플레이어의 점수가 딜러보다 낮은 경우 패배 반환")
+        void comparePlayerIfPlayerScoreSmallerThanDealer() {
+            final Dealer dealer = new Dealer(new CardGroup(cardKing, cardKing));
+            final Player player = new Player("홍실", initialGroup);
+
+            assertThat(player.calculatePlayerWinningStatus(dealer)).isEqualTo(WinningStatus.LOSE);
+        }
+
+        @Test
+        @DisplayName("딜러는 버스트가 아니고, 플레이어는 버스트인 경우 패배 반환")
+        void comparePlayerIfDealerNonBustPlayerBust() {
+            final Player player = new Player("홍실", initialGroup);
+            final Dealer dealer = new Dealer(new CardGroup(cardKing, cardEight));
+
+            player.drawCard(new Deck(new TestDeckGenerator(List.of(cardEight))));
+
+            assertThat(player.calculatePlayerWinningStatus(dealer)).isEqualTo(WinningStatus.LOSE);
+        }
     }
 }
